requirements.txt
mem0ai>=1.0.10
ollama>=0.6.0
fastmcp>=3.2.0
